+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 ### Added
 - Added method for adding piecewise linear constraints
 - Add SCIP function SCIPgetTreesizeEstimation and wrapper getTreesizeEstimation
-- Add recipes sub-package
+- New test for model setLogFile
 ### Fixed
+- Fixed model.setLogFile(None) error
+- Add recipes sub-package
 - Fixed "weakly-referenced object no longer exists" when calling dropEvent in test_customizedbenders
 - Fixed incorrect writing/printing when user had a non-default locale
 ### Changed